requirements/dev.txt
#
# This file is autogenerated by pip-compile
# Make changes in dev.in, then run this to update:
#
#    pip-compile dev.in
#
-e git+git://github.com/ariscn/pygerduty.git#egg=pygerduty
appnope==0.1.0            # via ipython
asq==1.2.1
blinker==1.4
cherrypy==3.8.1
click==6.2                # via pip-tools
decorator==4.0.6          # via ipython, traitlets
first==2.0.1              # via pip-tools
flask-assets==0.11
flask-debugtoolbar==0.10.0
flask-script==2.0.5
flask==0.10.1
gnureadline==6.3.3        # via ipython
ipython-genutils==0.1.0   # via traitlets
ipython==4.0.1
itsdangerous==0.24
jinja2==2.8
mako==1.0.3
markupsafe==0.23
path.py==8.1.2            # via pickleshare
peewee==2.7.4
pexpect==4.0.1            # via ipython
pickleshare==0.5          # via ipython
pip-tools==1.4.2
ptyprocess==0.5           # via pexpect
python-consul==0.4.7
requests==2.8.1
simplegeneric==0.8.1      # via ipython
six==1.10.0
slacker==0.8.6
traitlets==4.0.0          # via ipython
webassets==0.11.1
werkzeug==0.11.2
